refactor: remove old entity permission value objects
