+++ app/api/core/application/specs/FileDelete.spec.ts
@@ -1,4 +1,4 @@
-import { EntityPermissionChecker } from '#api/core/domain/entity/EntityPermissionChecker.js';
+import { EntityPermissionChecker } from '#api/core/domain/entityAccessPolicy/EntityPermissionChecker.js';
 import { EntitiesDataSourceFactory } from '#api/core/infrastructure/factories/EntitiesDataSourceFactory.js';
 import { FilesDataSourceFactory } from '#api/core/infrastructure/factories/FilesDataSourceFactory.js';
 import { FilesServiceFactory } from '#api/core/infrastructure/factories/FilesServiceFactory.js';
