+++ app/api/core/application/specs/UpdateEntity.spec.ts
@@ -282,8 +282,6 @@ describe('UpdateEntityUseCase', () => {
             type: 'Empty',
           },
           obsoleteMetadata: [],
-          permissions: [],
-          published: false,
         },
         {
           sharedId: 'full_entity',
@@ -339,8 +337,6 @@ describe('UpdateEntityUseCase', () => {
             type: 'Empty',
           },
           obsoleteMetadata: [],
-          permissions: [],
-          published: false,
         },
       ]);
     });
