--- app/api/core/domain/entity/AccessLevel.ts
@@ -1,7 +1,0 @@
-enum AccessLevel {
-  Read = 'read',
-  Write = 'write',
-  Mixed = 'mixed',
-}
-
-export { AccessLevel };
