--- app/api/core/domain/entity/PermissionType.ts
@@ -1,7 +1,0 @@
-enum PermissionType {
-  User = 'user',
-  Group = 'group',
-  Public = 'public',
-}
-
-export { PermissionType };
