+++ app/api/core/infrastructure/mongodb/entity/MongoEntityPermissionChecker.ts
@@ -1,7 +1,7 @@
 import {
   EntityPermissionChecker,
   Specification,
-} from '#api/core/domain/entity/EntityPermissionChecker.js';
+} from '#api/core/domain/entityAccessPolicy/EntityPermissionChecker.js';
 import { EntityDBO } from '#api/entities.v2/database/schemas/EntityTypes.js';
 import { Result, ResultType } from '#api/core/libs/Result.js';
 import { User } from '#api/users.v2/model/User.js';
